色設定方法の統一
