+++ frontend/src/app/happiness/all/page.tsx
@@ -8,7 +8,8 @@ import { PeriodType } from '@/types/period'
 import { MessageType } from '@/types/message-type'
 import { ResponsiveContainer } from 'recharts'
 const MapSet = dynamic(() => import('@/components/map/mapset'), { ssr: false })
-import { GetPin, COLORS } from '@/components/utils/pin'
+import { GetPin } from '@/components/utils/pin'
+import { graphColors } from '@/theme/color'
 import {
   DateTimeTextbox,
   useDateTimeProps,
@@ -140,7 +141,7 @@ const HappinessAll: React.FC = () => {
           <ResponsiveContainer width="100%" height={300}>
             <LineGraph
               plotdata={OurHappiness[period]}
-              color={COLORS}
+              color={graphColors}
               xTickFormatter={renderCustomDayTick}
             />
           </ResponsiveContainer>
