+++ frontend/src/theme/color.ts
@@ -1,0 +1,83 @@
+export const mapColors = {
+  "RED": [
+    '#ff0000',
+    '#ff7070',
+    '#ff6666',
+    '#ff5656',
+    '#ff4c4c',
+    '#ff4242',
+    '#ff3838',
+    '#ff2d2d',
+    '#ff2323',
+    '#ff1919',
+  ],
+  "BLUE": [
+    '#007fff',
+    '#70b7ff',
+    '#66b2ff',
+    '#5badff',
+    '#51a8ff',
+    '#47a3ff',
+    '#3d9eff',
+    '#3399ff',
+    '#2893ff',
+    '#1e8eff',
+  ],
+  "GREEN": [
+    '#4BA724',
+    '#3DCB00',
+    '#3EC704',
+    '#40C308',
+    '#42BF0C',
+    '#43BB10',
+    '#45B714',
+    '#46B318',
+    '#48AF1C',
+    '#4AAB20',
+  ],
+  "YELLOW": [
+    '#FF00D8',
+    '#FFCCF7',
+    '#FC9BEE',
+    '#FF99EF',
+    '#F272DF',
+    '#F56FE1',
+    '#F86CE3',
+    '#FB69E5',
+    '#FF65E8',
+    '#FF32E0',
+  ],
+  "ORANGE": [
+    '#ff7f00',
+    '#ffb770',
+    '#ffb266',
+    '#ffab5b',
+    '#ffa851',
+    '#ffa347',
+    '#ff9e3d',
+    '#ff9933',
+    '#ff9328',
+    '#ff8e1e',
+  ],
+  "VIOLET": [
+    '#7f00ff',
+    '#b770ff',
+    '#b266ff',
+    '#ad5bff',
+    '#a851ff',
+    '#a347ff',
+    '#9e3dff',
+    '#9933ff',
+    '#9328ff',
+    '#8e1eff',
+  ],
+}
+
+export const graphColors = [
+  mapColors["BLUE"][0],//'#007fff'
+  mapColors["GREEN"][0],//'#4BA724',
+  mapColors["VIOLET"][0],//'#7f00ff',
+  mapColors["YELLOW"][0],//'#FF00D8',
+  mapColors["ORANGE"][0],//'#ff7f00',
+  mapColors["RED"][0],//'#ff0000',
+]
